correct favorites status on first load try to fix
